refactor: create custom hook to force logout of user
